Number of reference populations was hard coded to 3
The grouper function was being called with it's n parameter always set to 3. This is a problem if you want to use 2 or 4+ reference populations. To fix this, I made a modification where the the n parameter is instead set to the number of populations specified in the command argument.